Add missing tests for Budget CRUD

+++ pkg/handler/budgets.go
@@ -4,9 +4,11 @@ import (
 	"encoding/json"
 	"net/http"
 	"strconv"
+	"strings"
 	"tracker/pkg/models"
 	"tracker/pkg/repository"
 	"tracker/pkg/utils"
+	errors "tracker/pkg/utils"
 )
 
 type BudgetHandler struct {
@@ -19,22 +21,26 @@ func NewBudgetHandler(repo repository.BudgetRepository, transactionsRepo reposit
 }
 
 func (h *BudgetHandler) GetBudget(w http.ResponseWriter, r *http.Request) {
-	id, _ := strconv.Atoi(r.PathValue("id"))
-	budget, err := h.repo.GetBudget(id)
-	transactions, _ := h.transactionsRepo.GetTransactionsForBudget(id)
 	encoder := json.NewEncoder(w)
+	parts := strings.Split(r.URL.Path, "/")
+	id, err := strconv.Atoi(parts[len(parts)-1])
+
+	if err != nil {
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
+		return
+	}
+
+	budget, err := h.repo.GetBudget(id)
+
+	if err == errors.Budget404Err {
+		utils.ErrRes(w, errors.Budget404Err, http.StatusNotFound)
+		return
+	}
+
+	transactions, _ := h.transactionsRepo.GetTransactionsForBudget(id)
 	budgetWithTransaction := models.BudgetWithTransactions{
 		Budget:       budget,
 		Transactions: transactions,
-	}
-
-	// TODO handle different type of error
-	// TODO write tests
-	if err != nil {
-		errPayload := utils.ErrRes(err)
-		w.WriteHeader(http.StatusNotFound)
-		encoder.Encode(errPayload)
-		return
 	}
 
 	w.WriteHeader(http.StatusOK)
@@ -45,12 +51,8 @@ func (h *BudgetHandler) GetBudgets(w http.ResponseWriter, r *http.Request) {
 	budgets, err := h.repo.GetBudgets()
 	encoder := json.NewEncoder(w)
 
-	// TODO handle different type of error
-	// TODO write tests
 	if err != nil {
-		errPayload := utils.ErrRes(err)
-		w.WriteHeader(http.StatusBadRequest)
-		encoder.Encode(errPayload)
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
 		return
 	}
 
@@ -63,14 +65,15 @@ func (h *BudgetHandler) CreateBudget(w http.ResponseWriter, r *http.Request) {
 	encoder := json.NewEncoder(w)
 
 	if err := json.NewDecoder(r.Body).Decode(&budget); err != nil {
-		errPayload := utils.ErrRes(err)
-		w.WriteHeader(http.StatusBadRequest)
-		encoder.Encode(errPayload)
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
 		return
 	}
 
-	// TODO handle errors
-	// TODO write tests
+	if budget.Title == "" {
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
+		return
+	}
+
 	payload := models.GenericPayload{Msg: "Budget created"}
 	h.repo.CreateBudget(budget)
 	w.WriteHeader(http.StatusCreated)
@@ -78,30 +81,53 @@ func (h *BudgetHandler) CreateBudget(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *BudgetHandler) DeleteBudget(w http.ResponseWriter, r *http.Request) {
-	id, _ := strconv.Atoi(r.PathValue("id"))
 	encoder := json.NewEncoder(w)
+	parts := strings.Split(r.URL.Path, "/")
+	id, err := strconv.Atoi(parts[len(parts)-1])
+
+	if err != nil {
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
+		return
+	}
+
+	if _, err := h.repo.GetBudget(id); err == errors.Budget404Err {
+		utils.ErrRes(w, errors.Budget404Err, http.StatusNotFound)
+		return
+	}
 
 	if err := h.repo.DeleteBudget(id); err != nil {
-		errPayload := utils.ErrRes(err)
-		w.WriteHeader(http.StatusBadRequest)
-		encoder.Encode(errPayload)
+		utils.ErrRes(w, err, 400)
 		return
 	}
 
 	payload := models.GenericPayload{Msg: "Budget deleted"}
-	w.WriteHeader(http.StatusAccepted)
+	w.WriteHeader(http.StatusNoContent)
 	encoder.Encode(payload)
 }
 
 func (h *BudgetHandler) UpdateBudget(w http.ResponseWriter, r *http.Request) {
 	var budget models.Budget
-	id, _ := strconv.Atoi(r.PathValue("id"))
+	parts := strings.Split(r.URL.Path, "/")
+	id, err := strconv.Atoi(parts[len(parts)-1])
 	encoder := json.NewEncoder(w)
 
+	if err != nil {
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
+		return
+	}
+
+	if _, err := h.repo.GetBudget(id); err == errors.Budget404Err {
+		utils.ErrRes(w, errors.Budget404Err, http.StatusNotFound)
+		return
+	}
+
 	if err := json.NewDecoder(r.Body).Decode(&budget); err != nil {
-		errPayload := utils.ErrRes(err)
-		w.WriteHeader(http.StatusBadRequest)
-		encoder.Encode(errPayload)
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
+		return
+	}
+
+	if budget.Title == "" {
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
 		return
 	}
